refactor eventHandlers to be within midiNotes component
